Improve activity page (#126)
* Update EditableFilteringInput to allow for a nullable onDeleteOption function

* Update ActivityScreen to have locations dropdown without delete option

* Remove unused method from ActivityViewModel

* Update EditableFilteringInput to be more responsive

+++ app/src/main/java/com/example/outdoorsy/ui/activity/components/EditableFilteringInput.kt
@@ -39,7 +39,7 @@ internal fun EditableFilteringInput(
     prompt: String,
     selectedText: String,
     onValueSelected: (String) -> Unit,
-    onDeleteOption: (String) -> Unit
+    onDeleteOption: ((String) -> Unit)? = null
 ) {
     var expanded by remember { mutableStateOf(false) }
     var text by remember { mutableStateOf(selectedText) }
@@ -89,7 +89,7 @@ internal fun EditableFilteringInput(
             )
 
             ExposedDropdownMenu(
-                expanded = expanded && (text.isNotBlank() || filteredOptions.isNotEmpty()),
+                expanded = expanded,
                 onDismissRequest = { expanded = false },
                 modifier = Modifier.clip(RoundedCornerShape(12.dp))
             ) {
@@ -112,13 +112,15 @@ internal fun EditableFilteringInput(
                                     horizontalArrangement = Arrangement.SpaceBetween
                                 ) {
                                     Text(option)
-                                    Icon(
-                                        imageVector = Icons.Default.Close,
-                                        contentDescription = "Delete Option",
-                                        modifier = Modifier.clickable {
-                                            onDeleteOption(option)
-                                        }
-                                    )
+                                    if (onDeleteOption != null) {
+                                        Icon(
+                                            imageVector = Icons.Default.Close,
+                                            contentDescription = "Delete Option",
+                                            modifier = Modifier.clickable {
+                                                onDeleteOption(option)
+                                            }
+                                        )
+                                    }
                                 }
                             },
                             onClick = {
